kms_service/requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
requests>=2.31.0
cryptography>=42.0.0
django-cors-headers==4.9.0
openpyxl==3.1.2
